backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.2.1
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.0
Deprecated==1.2.18
durationpy==0.9
fastapi==0.115.8
filelock==3.17.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==25.2.10
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.67.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kubernetes==32.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.0.0
oauthlib==3.2.2
onnxruntime==1.20.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pillow==11.1.0
posthog==3.13.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.4.1
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.3
wrapt==1.17.2
zipp==3.21.0
